Allow Non-Pixel-Perfect Sprites
Major rendering update that allows for non-pixel-perfect sprites while
still forcing sprites to be perfectly aligned by default.

Collisions were removed temporarily. A new collision system will be
added in another commit.

+++ crates/bevy_retrograde_ldtk/src/system.rs
@@ -157,7 +157,11 @@ fn process_ldtk_maps(
                             },
                             // Each layer is 2 units higher than the one before it
                             visible: Visible(layer.visible),
-                            position: Position::new(level.world_x, level.world_y, z as i32 * 2),
+                            position: Position::new(
+                                level.world_x as f32,
+                                level.world_y as f32,
+                                z as f32 * 2.,
+                            ),
                             ..Default::default()
                         })
                         .insert(LdtkMapLayer {
